% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_unique_id.R
\name{make_unique_id}
\alias{make_unique_id}
\title{Create a unique ID for naming files}
\usage{
make_unique_id(safe = TRUE)
}
\arguments{
\item{safe}{Logical value - do not edit}
}
\value{
code{make_unique_id} returns a unique ID name
}
\description{
\code{make_unique_id} is used purely to generate unique file names for exported tables.
}
\examples{
\dontrun{
make_unique_id()
[1] "5d31f2bb"
}

}
