% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_data.R
\name{record_data}
\alias{record_data}
\title{Record four-arm olfactometer data}
\usage{
record_data(x)
}
\arguments{
\item{x}{Leave blank}
}
\value{

}
\description{
Create a .txt file containing the duration of time a study subject spends
in each of the five olfactometer zones and the number of times each zone
was entered. When this function is called the user will be prompted in the
console to enter experimental details required for naming the .txt file and
for further analysis.
}
\examples{
\dontrun{
record_data()

User initials: JR
Year: 2019
Experiment number: 1
Replicate number: 1
Olfactometer arm containing treatment: 2
Press s to begin recording data: s
Olfactometer zone: 5
Olfactometer zone: 4
13.7 sec elapsed
Olfactometer zone: 2
6.19 sec elapsed
Olfactometer zone: 5
6.61 sec elapsed
Olfactometer zone: 2
45.67 sec elapsed
Olfactometer zone: t

}
}
