% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_four_arm.R
\name{record_four_arm}
\alias{record_four_arm}
\title{Record movement data from four-arm olfactometer experiments}
\usage{
record_four_arm()
}
\value{
\code{record_four_arm} returns a summary table of a study subject's movements
recorded in a four-arm olfactometer to the console and exports a .xlsx file containing
these values to the user's working directory.
}
\description{
\code{record_four_arm} allows the user to interactively record the movements
of their study subject in a four-arm olfactometer.

Upon executing \code{record_four_arm} the user will be prompted in the console to enter the
following information before recording: User initials', 'Year of experiment',
'Experiment number', 'Replicate number' and 'Olfactometer arm containing treatment'.

Four-arm olfactometers are typically divided into five unique zones (one for each arm
as well as a central zone), with each zone corresponding to a numerical
key from one to five. When a study subject leaves a zone, the user must use the
numerical key corresponding to the departed zone to record the correct zone. When the
observation period is over, the user can terminate the recording process by pressing 't'.
}
\examples{
\dontrun{> library(olfactometeR)
> record_four_arm()
User initials: JR
Year: 2019
Experiment number: 1
Replicate number: 1
Olfactometer arm containing treatment: 2
Press any key to begin recording data:
Olfactometer zone: 5
98.08 sec elapsed
Olfactometer zone: 4
25.61 sec elapsed
Olfactometer zone: 2
106.95 sec elapsed
Olfactometer zone: 4
3.57 sec elapsed
Olfactometer zone: 2
11.85 sec elapsed
Olfactometer zone: 5
11.37 sec elapsed
Olfactometer zone: 3
55.02 sec elapsed
Olfactometer zone: 2
11.72 sec elapsed
Olfactometer zone: t


|Olfactometer Zone|Time in Zone (secs)|Time in Zone (mins)|No. Times Zone Entered|Treatment Arm|
|:---------------:|:-----------------:|:-----------------:|:--------------------:|:-----------:|
|         1       |        0.00       |       0.00        |          0           |             |
|         2       |      130.52       |       2.18        |          3           |      T      |
|         3       |       55.02       |       0.92        |          1           |             |
|         4       |       29.18       |       0.49        |          2           |             |
|         5       |      109.45       |       1.82        |          2           |             |
}

}
