% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_six_arm.R
\name{record_six_arm}
\alias{record_six_arm}
\title{Record behavioural data from six-arm olfactometer experiments}
\usage{
record_six_arm()
}
\value{
\code{record_six_arm} returns a summary table of a study subject's movements
recorded in a six-arm olfactometer to the console.
}
\description{
\code{record_six_arm} allows the user to interactively record the movements
of their study subject in a six-arm olfactometer.

Upon executing \code{record_six_arm} the user will be prompted in the console to enter the
following information before recording: User initials', 'Year of experiment',
'Experiment number', 'Replicate number', 'Centre zone assignment', 'Number of
treatment arms' and 'Olfactometer arm containing treatment'.

Six-arm olfactometers are typically divided into seven unique zones (one for each arm
as well as a central zone), with each zone corresponding to a numerical
key from one to seven. When a study subject leaves a zone, the user must use the
numerical key corresponding to the departed zone to record the correct zone. When the
observation period is over, the user can terminate the recording process by pressing 't'.
}
\examples{
\dontrun{

# One treatment arm

Library(olfactometeR)

record_six_arm()
User initials: JR
Year: 2019
Experiment number: 1
Replicate number: 1
Centre zone (1/2/3/4/5/6/7): 7
Number of treatment arms (1/2): 1
Olfactometer arm containing treatment (1/2/3/4/5/6/7): 1
Press any key to begin recording data:
Olfactometer zone: 7
17.58 sec elapsed
Olfactometer zone: 4
17.39 sec elapsed
Olfactometer zone: 7
20.77 sec elapsed
Olfactometer zone: 6
27.27 sec elapsed
Olfactometer zone: 7
6.44 sec elapsed
Olfactometer zone: 1
39.47 sec elapsed
Olfactometer zone: t
|Olfactometer Zone|Time in Zone (secs)|Time in Zone (mins)|No. Times Zone Entered|Zone Assignment|
|:---------------:|:-----------------:|:-----------------:|:--------------------:|:-------------:|
|        1        |       39.47       |        0.66       |         1            |   Treatment   |
|        2        |        0.00       |        0.00       |         0            |    Control    |
|        3        |        0.00       |        0.00       |         0            |    Control    |
|        4        |       17.39       |        0.29       |         1            |    Control    |
|        5        |        0.00       |        0.00       |         0            |    Control    |
|        6        |       27.27       |        0.45       |         1            |    Control    |
|        7        |       44.79       |        0.75       |         3            |    Centre     |

}
}
