% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_y_tube.R
\name{record_y_tube}
\alias{record_y_tube}
\title{Record choice data from Y-tube olfactometer experiments}
\usage{
record_y_tube()
}
\value{
code{record_y_tube} returns a summary table in the console.
}
\description{
\code{record_y_tube} allows the user to record the choices made by
study subjects in Y-tube olfactometer experiments.

Upon executing \code{record_y_tube} the user will be prompted in the console to enter the
following information before recording: User initials', 'Year of experiment',
'Experiment number', 'Replicate number' and 'Olfactometer arm containing treatment'.

Each Y-tube olfactometer arm corresponds to a numerical key, either one or two. When
a study subject enters an olfactometer arm and crosses the pre-determined
decision line the user must use the numerical key corresponding to the olfactometer arm to
record the individual as making a choice. Recording will automatically terminate once
a choice has been made.
}
\examples{
\dontrun{> library(olfactometeR)
record_y_tube()
User initials: JR
Year: 2019
Experiment number: 1
Replicate number: 1
Olfactometer arm containing treatment (1:2): 2
Press any key to begin recording data:
Olfactometer zone: 1
99.97 sec elapsed
|Olfactometer Arm|Time to Reach Arm End (secs)|Time to Reach Arm End (mins)|Treatment Arm|
|:--------------:|:--------------------------:|:--------------------------:|:-----------:|
|        1       |            99.97           |             1.67           |             |
|        2       |              NA            |              NA            |      T      |
}

}
