% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_four_arm.R
\name{summarise_four_arm}
\alias{summarise_four_arm}
\title{Summarise four-arm olfactometer experiment results}
\usage{
summarise_four_arm()
}
\value{
\code{summarise_four_arm} returns a summary table of the experiment
replicates to the console, which can be exported as a .xlsx file if required.
}
\description{
\code{summarise_four_arm} allows the user to summarise replicates
from four-arm olfactometer experiments.

Upon executing \code{summarise_four_arm} the user will be prompted to select the
experiment replicates they want to summarise, displaying a summary table in the
console that can be exported as a .xlsx file if required.
}
\examples{
\dontrun{
library(olfactometeR)
summarise_four_arm()
Number of treatment arms (1/2): 1


| Replicate | Centre | Treatment | Control 1 | Control 2 | Control 3 | Control mean |
|:---------:|:------:|:---------:|:---------:|:---------:|:---------:|:------------:|
|     1     | 72.52  |   4.83    |   0.00    |   90.47   |   4.20    |    31.56     |
|     2     | 20.06  |   3.98    |   0.00    |   21.71   |  366.61   |    129.44    |
|     3     | 62.03  |   1.32    |  129.27   |   79.26   |   0.00    |    69.51     |

Save the ouput as an .xlsx file? (y/n) n


|                          |
|:-------------------------|
|Output has not been saved |
}

}
