% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_y_tube.R
\name{summarise_y_tube}
\alias{summarise_y_tube}
\title{Summarise Y-tube olfactometer experiment results}
\usage{
summarise_y_tube()
}
\value{
\code{summarise_y_tube} returns a summary table of the experiment
replicates to the console, which can be exported as a .xlsx file if required.
}
\description{
\code{summarise_y_tube} allows the user to summarise replicates
from Y-tube olfactometer experiments.

Upon executing \code{summarise_y_tube} the user will be prompted to select the
experiment replicates they want to summarise, displaying a summary table in the
console that can be exported as a .xlsx file if required.
}
\examples{
\dontrun{
library(olfactometeR)

summarise_y_tube()

Save the ouput as an .xlsx file? (y/n) n

[1] "Output has not been saved"
}

}
