% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_data.R
\name{record_data}
\alias{record_data}
\title{Record four-arm olfactometer data}
\usage{
record_data(x)
}
\arguments{
\item{x}{}
}
\description{
Create a .txt file containing the duration of time a study subject spends
in each of the five olfactometer zones and the number of times each zone
was entered. When this function is called the user will be prompted in the
console to enter experimental details required for naming the .txt file and
for further analysis.
}
\examples{
record_data()
}
