% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_four_arm.R
\name{record_four_arm}
\alias{record_four_arm}
\title{\code{record_four_arm} - Records data from four-arm olfactometers}
\usage{
record_four_arm(x)
}
\arguments{
\item{x}{Leave blank}
}
\value{
\code{record_four_arm} allows the user to interactively record an
individual's movements in a four-arm olfactometer, returing a summary
table in the console and exporting an .xlsx file containing the results.
}
\description{
\code{record_four_arm} - Records data from four-arm olfactometers
}
\examples{
\dontrun{record_four_arm()

User initials: JR
Year: 2019
Experiment number: 1
Replicate number: 1
Olfactometer arm containing treatment: 2
Press any key to begin recording data: s
Olfactometer zone: 5
Olfactometer zone: 4
13.7 sec elapsed
Olfactometer zone: 2
6.19 sec elapsed
Olfactometer zone: 5
6.61 sec elapsed
Olfactometer zone: 2
45.67 sec elapsed
Olfactometer zone: t
}

}
