% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_four_arm.R
\name{record_four_arm}
\alias{record_four_arm}
\title{Record behavioural data from four-arm olfactometer experiments}
\usage{
record_four_arm()
}
\value{
\code{record_four_arm} returns a summary table of a study subject's movements
recorded in a four-arm olfactometer to the console.
}
\description{
\code{record_four_arm} allows the user to interactively record the movements
of their study subject in a four-arm olfactometer.

Upon executing \code{record_four_arm} the user will be prompted in the console to enter the
following information before recording: User initials', 'Year of experiment',
'Experiment number', 'Replicate number', 'Centre zone assignment', 'Number
of treatment arms' and Olfactometer arm containing treatment'.

Four-arm olfactometers are typically divided into five unique zones (one for each arm
as well as a central zone), with each zone corresponding to a numerical
key from one to five. When a study subject leaves a zone, the user must use the
numerical key corresponding to the departed zone to record the correct zone. When the
observation period is over, the user can terminate the recording process by pressing 't'.
}
\examples{
\dontrun{

# One treatment arm

Library(olfactometeR)

record_four_arm()
User initials: JR
Year: 2019
Experiment number: 1
Replicate number: 1
Number of treatment arms (1/2): 1
Olfactometer arm containing treatment (1/2/3/4): 2
Press any key to begin collecting data:
Olfactometer zone: 5
98.08 sec elapsed
Olfactometer zone: 4
25.61 sec elapsed
Olfactometer zone: 2
106.95 sec elapsed
Olfactometer zone: 4
3.57 sec elapsed
Olfactometer zone: 2
11.85 sec elapsed
Olfactometer zone: 5
11.37 sec elapsed
Olfactometer zone: 3
55.02 sec elapsed
Olfactometer zone: 2
11.72 sec elapsed
Olfactometer zone: t


|Olfactometer Zone|Time in Zone (secs)|Time in Zone (mins)|No. Times Zone Entered|Treatment Arm|
|:---------------:|:-----------------:|:-----------------:|:--------------------:|:-----------:|
|        1        |        0.00       |       0.00        |          0           |             |
|        2        |      130.52       |       2.18        |          3           |      T      |
|        3        |       55.02       |       0.92        |          1           |             |
|        4        |       29.18       |       0.49        |          2           |             |
|        5        |      109.45       |       1.82        |          2           |             |


# Two treatment arms

library(olfactometeR)

record_four_arm()
User initials: JR
Year: 2019
Experiment number: 1
Replicate number: 1
Number of treatment arms (1/2): 2
Olfactometer arm containing treatment one (1/2/3/4): 1
Olfactometer arm containing treatment two (1/2/3/4): 4
Press any key to begin collecting data:
Olfactometer zone: 5
33.94 sec elapsed
Olfactometer zone: 4
23.4 sec elapsed
Olfactometer zone: 5
5.91 sec elapsed
Olfactometer zone: 1
42.33 sec elapsed
Olfactometer zone: 5
28.64 sec elapsed
Olfactometer zone: 3
8.36 sec elapsed
Olfactometer zone: t


|Olfactometer Zone|Time in Zone (secs)|Time in Zone (mins)|No. Times Zone Entered|Treatment Arms|
|:---------------:|:-----------------:|:-----------------:|:--------------------:|:------------:|
|        1        |       42.33       |        0.71       |           1          |       T      |
|        2        |        0.00       |        0.00       |           0          |              |
|        3        |        8.36       |        0.14       |           1          |              |
|        4        |       23.40       |        0.39       |           1          |       T      |
|        5        |       68.49       |        1.14       |           3          |              |

}

}
