% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_y_tube.R
\name{record_y_tube}
\alias{record_y_tube}
\title{Record choice data from Y-tube olfactometer experiments}
\usage{
record_y_tube()
}
\value{
code{record_y_tube} returns a summary table in the console.
}
\description{
\code{record_y_tube} allows the user to record the choices made by
study subjects in Y-tube olfactometer experiments.

Upon executing \code{record_y_tube} the user will be prompted in the console to enter the
following information before recording: User initials', 'Year of experiment',
'Experiment number', 'Replicate number' and 'Olfactometer arm containing treatment'.

Each Y-tube olfactometer arm corresponds to a numerical key, either one or two. When
a study subject enters an olfactometer arm and crosses the pre-determined
decision line the user must use the numerical key corresponding to the olfactometer arm to
record the individual as making a choice. Recording will automatically terminate once
a choice has been made.
}
\examples{
\dontrun{
library(olfactometeR)
record_y_tube()
User initials: JR
Year: 2019
Study species being testsed: "Phytoseiulus persimilis"
Experiment number: 1
Replicate number: 1
Olfactometer arm containing treatment (1/2): 2
Treatment: "Linalool"


 |Variable                 |User Response             |
 |:------------------------|:-------------------------|
 |User initials            |JR                        |
 |Study year               |2030                      |
 |Study subject species    |"Phytoseiulus persimilis" |
 |Experiment no.           |1                         |
 |Replicate no.            |1                         |
 |Treatment arm assignment |2                         |
 |Treatment                |"linalool"                |

 Are the entered details correct (y/n): y
 Press any key to begin recording data:
 Olfactometer zone: 1
 29.2 sec elapsed


 |Olfactometer Arm|Zone Assignment|Time to Reach Arm End (secs)|Time to Reach Arm End (mins)|
 |:--------------:|:-------------:|:--------------------------:|:--------------------------:|
 |       1        |    Control    |           29.2             |           0.49             |
 |       2        |   Treatment   |            NA              |            NA              |
 Save the ouput (y/n): n
 [1] "Output has not been saved"
}

}
