% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_y_tube.R
\name{record_y_tube}
\alias{record_y_tube}
\title{\code{record_y_tube} - Records data from Y-tube olfactometers}
\usage{
record_y_tube(x)
}
\arguments{
\item{x}{Leave blank}
}
\value{
code{record_y_tube} allows the user to interactively record the
choice of an individual in a Y-tubeolfactometer, returing a summary
table in the console and exporting an .xlsx file containing the result.
}
\description{
\code{record_y_tube} - Records data from Y-tube olfactometers
}
\examples{
\dontrun{> library(olfactometeR)
record_y_tube()
User initials: JR
Year: 2019
Experiment number: 1
Replicate number: 1
Olfactometer arm containing treatment: 2
Press any key to begin recording data:
Olfactometer zone: 1
99.97 sec elapsed
|Olfactometer Arm|Time to Reach Arm End (secs)|Time to Reach Arm End (mins)|Treatment Arm|
|:--------------:|:--------------------------:|:--------------------------:|:-----------:|
|        1       |            99.97           |             1.67           |             |
|        2       |              NA            |              NA            |      T      |
}

}
