% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_y_tube.R
\name{record_y_tube}
\alias{record_y_tube}
\title{\code{record_y_tube} - Records data from Y-tube olfactometers}
\usage{
record_y_tube(x)
}
\arguments{
\item{x}{Leave blank}
}
\value{
code{record_y_tube} allows the user to interactively record the
choice of an individual in a Y-tubeolfactometer, returing a summary
table in the console and exporting an .xlsx file containing the result.
}
\description{
\code{record_y_tube} - Records data from Y-tube olfactometers
}
\examples{
\dontrun{> library(olfactometeR)
}

}
