% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_y_tube.R
\name{record_y_tube}
\alias{record_y_tube}
\title{Record choice data from Y-tube olfactometer experiments}
\usage{
record_y_tube()
}
\value{
code{record_y_tube} returns a summary table in the console.
}
\description{
\code{record_y_tube} allows the user to record the choices made by
study subjects in Y-tube olfactometer experiments.

Upon executing \code{record_y_tube} the user will be prompted in the console to enter the
following information before recording: User initials', 'Year of experiment',
'Experiment number', 'Replicate number' and 'Olfactometer arm containing treatment'.

Each Y-tube olfactometer arm corresponds to a numerical key, either one or two. When
a study subject enters an olfactometer arm and crosses the pre-determined
decision line the user must use the numerical key corresponding to the olfactometer arm to
record the individual as making a choice. Recording will automatically terminate once
a choice has been made.
}
\examples{
\dontrun{
library(olfactometeR)

record_y_tube()

User initials: JR
Year: 2019
Study species being tested: "Myzus"                ## Must be entered between quotation marks!
Experiment number: 1
Replicate number: 2
Olfactometer arm containing treatment (1/2): 2
Treatment: "Limonene (0.1 mg/ml)"                   ## Must be entered between quotation marks!


|Variable                 |User Response          |
|:------------------------|:----------------------|
|User initials            |JR                     |
|Study year               |2019                   |
|Study subject species    |"Myzus"                |
|Experiment no.           |1                      |
|Replicate no.            |2                      |
|Treatment arm assignment |2                      |
|Treatment                |"Limonene (0.1 mg/ml)" |

Are the entered details correct (y/n): y

Press any key to begin recording data:
Olfactometer zone: 2
3.33 sec elapsed

                                  Y-Tube olfactometer
  ------------------------------------------------------------------------------------
  Olfactometer Arm   Zone Assignment   Time to Arm End (secs)   Time to Arm End (mins)
  ------------------------------------------------------------------------------------
         1               Control
         2              Treatment               3.33                     0.06
  ------------------------------------------------------------------------------------
  Study species: Myzus persicae
  Treatment: Limonene (0.1 mg/ml)

Save the ouput (y/n): n
[1] "Output has not been saved"
}

}
