% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_four_arm.R
\name{summarise_four_arm}
\alias{summarise_four_arm}
\title{Summarise four-arm olfactometer experiment results}
\usage{
summarise_four_arm()
}
\value{
\code{summarise_four_arm} returns a summary table of the experiment
replicates to the console, which can be exported as a .xlsx file if required.
}
\description{
\code{summarise_four_arm} allows the user to summarise replicates
from four-arm olfactometer experiments.

Upon executing \code{summarise_four_arm} the user will be prompted to select the
experiment replicates they want to summarise. displaying a summary table in the
console that can be exported as a .xlsx file if required.
}
\examples{
\dontrun{
Example goes here.
}

}
